updated v2 (buggy)
still buggy figuring out the game mechanics now. Seems as if there is a glitch if you enter anything other than what is in the hashmap!

Fixed the bug where the hashmap arraylist value changes because it turns out its a pointer so i just created a 'new' variable in its place.